Added some tests. Made token literals easier to differentiate from their values. Upgraded to c++20; using g++-9 now.

+++ src/scanner.cpp
@@ -139,7 +139,7 @@ namespace Scan
      */
     bool Scanner::isAtEnd()
     {
-        return current >= source.length();
+        return current >= (int)source.length();
     }
 
     /**
